Hey Symphonians,
I was looking to do a community Bible-based church (we're calvary chapel) website project. One that has people involved and can be distributed freely. I've chatted with several other Symphony users who would be interested, but just wanted to see if there were others. It would be a large code-share and basically I want the project to benefit from all active participants.

We would run the project management thru basecamp, and chat thru campfire ( both by 37 Signals )

The ideal first deliverable is something that manages members, events/classes (including commerce [after first release]), pages, teachings, volunteering, ustream API, etc. Basically a way that any church can download this ensemble and install Church website 1.0 and be fully functional on day one.

I'd like the project to be on a shared repository so the code is constantly updated. The hope is that we make this a community project so no one-person is responsible, but instead has a growing user base of content contributors. Ideally someone could install the website with a church that didn't have internal staff resources for a website, and that the site could be easily updatable in the future (auto-updatedable like WordPress installs would be awesome)

Cheers !